# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: ESG & Environmental Data Management Platform API
  description: |
    REST API for the ESG Environmental Data Management Platform.
    Manages energy, water, waste, GHG emissions, air emissions, sales, recovery,
    and production data across manufacturing sites.
  version: 2026-02-24
  contact:
    name: ESG Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/2026-02-24
    description: Local development
  - url: http://localhost:3000/api/2026-02-24
    description: Docker (via nginx proxy)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    SiteRef:
      type: object
      properties:
        name:
          type: string
        code:
          type: string

    # ── Data Entry Common ──
...
    ProductionData:
      type: object
      required: [siteId, month, year, quantityMt]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        quantityMt:
          type: number
          minimum: 0
        site:
          $ref: '#/components/schemas/SiteRef'
...
    ETPData:
      type: object
      required: [siteId, month, year]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        codMgPerL:
          type: number
          minimum: 0
        bodMgPerL:
          type: number
          minimum: 0
        tssMgPerL:
          type: number
          minimum: 0
        tdsMgPerL:
          type: number
          minimum: 0
        sludgeMt:
          type: number
          minimum: 0
        site:
          $ref: '#/components/schemas/SiteRef'
    GHGEmission:
      type: object
      required: [siteId, month, year, scope, sourceDescription, activityData, activityUnitId, emissionFactorId, co2eTonnes]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        scope:
          type: string
          enum: [scope_1, scope_2, scope_3]
        scope1Category:
          type: string
          enum: [stationary, mobile, process, fugitive]
        scope2Method:
          type: string
          enum: [location_based, market_based]
        scope3Category:
          type: string
          enum: [purchased_goods, capital_goods, fuel_energy, upstream_transport, waste_operations, business_travel, employee_commuting, upstream_leased, downstream_transport, processing_sold, use_of_sold, end_of_life, downstream_leased, franchises, investments]
        sourceDescription:
          type: string
          maxLength: 255
        activityData:
          type: number
          minimum: 0
        activityUnitId:
          type: integer
        emissionFactorId:
          type: integer
        co2Tonnes:
          type: number
        ch4Tonnes:
          type: number
        n2oTonnes:
          type: number
        co2eTonnes:
          type: number
          minimum: 0
        site:
          $ref: '#/components/schemas/SiteRef'
...
    RecoveryData:
      type: object
      required: [siteId, month, year, materialName, quantityMt]
      properties:
        id:
          type: integer
          readOnly: true
        siteId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        materialName:
          type: string
          maxLength: 200
        quantityMt:
          type: number
          minimum: 0
        revenueLkr:
          type: number
          minimum: 0
        site:
          $ref: '#/components/schemas/SiteRef'
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
security:
  - bearerAuth: []
paths:
...
  /data/production:
    get:
      tags: [Production]
      summary: List production data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of production records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductionData'
    post:
      tags: [Production]
      summary: Create production entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionData'
      responses:
        '201':
          description: Created
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Water ──
...
  /data/etp:
    get:
      tags: [ETP]
      summary: List ETP data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of ETP records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ETPData'
    post:
      tags: [ETP]
      summary: Create ETP entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ETPData'
      responses:
        '201':
          description: Created
  /data/etp/{id}:
    put:
      tags: [ETP]
      summary: Update ETP entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [ETP]
      summary: Delete ETP entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── GHG ──
  /data/ghg:
    get:
      tags: [GHG]
      summary: List GHG emissions data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: scope
          in: query
          schema:
            type: string
            enum: [scope_1, scope_2, scope_3]
      responses:
        '200':
          description: Array of GHG records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GHGEmission'
    post:
      tags: [GHG]
      summary: Create GHG emission entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GHGEmission'
      responses:
        '201':
          description: Created
  /data/ghg/{id}:
    put:
      tags: [GHG]
      summary: Update GHG entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [GHG]
      summary: Delete GHG entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Air Emissions ──
...
  /data/recovery:
    get:
      tags: [Recovery]
      summary: List recovery data
      parameters:
        - name: siteId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Array of recovery records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecoveryData'
    post:
      tags: [Recovery]
      summary: Create recovery entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoveryData'
      responses:
        '201':
          description: Created
  /data/recovery/{id}:
    put:
      tags: [Recovery]
      summary: Update recovery entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Recovery]
      summary: Delete recovery entry (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Dashboard ──
...